Add Asteroid class and make DISPLAY_SIZE global
Implemented initialization of all asteroids along with a difficulty
mechanism for increasing the number that spawn. Implemented asteroid
motion and logic to reset the asteroids when the game resets.

Made DISPLAY_SIZE a global const in the ag namespace because it was
becoming hard to keep up with all the places it needed to be passed
around.