Move ship sprite management to Spaceship class
Spaceship class now owns and updates its own sprite.

In implementing this change a bug came up because the font object was
being passed by value and then getting destroyed before the text could
be rendered to the screen. In light of this all class objects are now
being passed to functions by const reference.

Also accompanying this commit is a bunch of cleanup. Removed a lot of
includes and functions/function prototypes that are no longer in use.
Privatized some class members that were made public for debugging.

Renamed some Action enum values in preparation for doing away with this
method of tracking user input.