Refactor user input handing and game update logic
Removed the unnecessary Action enum and refactored the logic so that
the effects of player input is processed in Game::process_input()
instead of Game::update(). This allows for better delineation between
the handling of input vs. the updating of the game state. Accompanied
this with a streamlining of the ship control logic.