Change the player spaceship to a ConvexShape
Player ship had been implemented as a CircleShape while following the
SFML tutorial. Re-implemented it as a ConvexShape because that makes
more sense. Also changed ship to an isosceles triangle for aesthetic
reasons. Due to the shape change, the collision is now no longer working
correctly and will need to be changed to an STA implementation. This
needed to be done anyway.