Change ship logic to only use sprite's data
As previously implemented, the player ship and the asteroids keep track
of position and orientation themselves but the SFML library shape object
that is used for their sprite is already storing those values. The ship
has now been refactored to use the existing SFML ConvexShape object
position and rotation members. The asteroids will be next.