Perform some minor cleanup
Move resources out of the source folder. Remove unnecessary value
initilizations for the sprites when the value they are initilizing to is
the same as the documented default value. Corrected an implicit cast
when calculating the starting position for the player spaceship. Moved
the declaration of dt outside of the main game loop.