Implement significant chaned to asteroid collision
Implemented a physics-based deflection calculation for the asteroids
that incorportaes mass and momentum transfer vectors. As a result the
deflections look a lot better. Screen wrapped deflections still result
in the asteroids getting stuc inside each other a lot of the time. May
need to spawn ghost asteroids when they screen wrap. Also refactored
collision detection code to be less of a mess.