Rework logic for screen wrapping asteroids
Screen wrapping asteroids by jumping them to the other side of the
viewport caused issues with collision between asteroids on opposite
edges of the screen. Replaced logic for warping asteroids to the other
side of the veiwport with logic that spawns a copy of an asteroid on the
opposite side of the viewport as the original moves off the screen. Bugs
remain. Namely that velocty between original and copy can become
different, and also that changes in velocity at the edges of the screen
can cause mutilple copies to be created.