Implement HUD and improve game state logic
Added logic to the DisplayManager to render lives and score in a HUD.
Also added text to the title screen.

Refactored state transition logic to move it more into the StateManager.
Also made transitions between game states work a bit different in order
to imporve the expecrience.

Removed debug code.